requirements.txt
Jinja2==2.10.1
azure-devops==5.0.0b9
msrest==0.6.8
flake8==3.7.8
pylint==2.3.1
